--- src/auth/decorator/customer.decorator.ts
@@ -1,9 +1,0 @@
-import { createParamDecorator, ExecutionContext } from '@nestjs/common';
-import { Customer } from '../../schemas/customers.entity';
-
-export const CurrentCustomer = createParamDecorator(
-  (data: unknown, context: ExecutionContext) => {
-    const req: { user?: Customer } = context.switchToHttp().getRequest();
-    return req.user;
-  },
-);
